removed commented code and old pages
