--- src/SFA.DAS.Registration.UITests/Project/Tests/Pages/YourTrainingProvidersPage.cs
@@ -1,34 +1,0 @@
-﻿using OpenQA.Selenium;
-using TechTalk.SpecFlow;
-
-namespace SFA.DAS.Registration.UITests.Project.Tests.Pages
-{
-    public class YourTrainingProvidersPage(ScenarioContext context) : RegistrationBasePage(context)
-    {
-        protected override string PageTitle => "Your training providers";
-        private static By AddANewTrainingProviderButton => By.LinkText("Add a training provider");
-        private static By SetPermissionsLink => By.PartialLinkText("Set permissions");
-
-        public EnterYourTrainingProviderNameReferenceNumberUKPRNPage SelectAddATrainingProvider()
-        {
-            formCompletionHelper.ClickElement(AddANewTrainingProviderButton);
-            return new EnterYourTrainingProviderNameReferenceNumberUKPRNPage(context);
-        }
-
-        public SetPermissionsForTrainingProviderPage SelectSetPermissions(string orgName)
-        {
-            if (string.IsNullOrEmpty(orgName))
-                formCompletionHelper.ClickElement(SetPermissionsLink);
-            else
-                tableRowHelper.SelectRowFromTable("Set permissions", orgName);
-
-            return new SetPermissionsForTrainingProviderPage(context);
-        }
-
-        public CreateYourEmployerAccountPage GoBackToCreateYourEmployerAccountPage()
-        {
-            formCompletionHelper.Click(BackLink);
-            return new CreateYourEmployerAccountPage(context);
-        }
-    }
-}
